refactor for better readability
